day15

+++ pkg/advent_helpers/int_helpers.go
@@ -6,7 +6,7 @@ func MinInt(a int, b int) int {
 	}
 	return b
 }
-func MaxInt(a int64, b int64) int64 {
+func MaxInt(a int, b int) int {
 	if a < b {
 		return b
 	}
